[FIX] sale_project: fix the name overwrite when creating an sol
This commit's purpose is to keep the user input when he creates an sol
from the formview sol.

Step to reproduce :

- open project update
- cilck on 'sols' stat button
- click on 'new' to create a new sol
- give a name 'youpi' to the sol
- save the sol

The name of the sol will be overwritten with a default one.
Expected behavior:
The name of the sol should be the one the user has set.

Source of the problem:
The name of sol is used for some computation in some use case when the
form view of sol from the sale_project module is used. But this form
view is also used in the use case where the name override is not needed.

Solution: add a context key to differentiate the 2 use cases.

task-4031977
version : 17.0 - master

closes odoo/odoo#171786

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>