[FIX] stock: raise a error message if product is not in stock move
This error occurs when we keep the product name empty in the stock move and try
to generate the Serials/Lots by clicking on 'Generate Serials/Lots'.

Steps to reproduce:
- Install the 'stock' module
- Create a new product(eg: test) and set its tracking 'By Lots' in Traceability
- Create a new receipt and add 'test' as product and 'Mark as ToDo'
- Again add 'test' as the product and on the right, click on 'bars'(fa-icon)
- Remove the product name > click on 'Generate Serials/Lots'
- Click on 'Generate'

Traceback: KeyError
'product_id'

This error occurs at [1] because 'product_id' is not present in the default
values.

This fix will display a dialog box with an error message if the product name is
empty in the stock move.

[1]: https://github.com/odoo/odoo/blob/b1ba4018f91f1fd788f48180bd5fc4656ab1e35c/addons/stock/models/stock_move.py#L930

sentry-5072366645

closes odoo/odoo#170718

Signed-off-by: Quentin Wolfs (quwo) <quwo@odoo.com>