[FIX] mrp_workorder: don't start a finished work order
Steps to reproduce the bug:
- Create a storable product P1 with BoM:
    - Component: Add any component
    - operation: OP1

- Create a MO with P1:
    - Confirm it
    - Start the workorder and mark it as done
- Come back to the dashboard >  workorders list
- Select the finished workorder
- Try to start it

Problem:
The work order can be started while it is in the 'done' state.

When the function `button_start` is called, we will check if we need to
skip the employee check or not. However, since the current user is also
an employee, we will use them:
https://github.com/odoo/enterprise/blob/c4604d8b398713898b014e6073219d6202c154c3/mrp_workorder/models/mrp_workorder.py#L258-L259
The function start_employee will then be called:
https://github.com/odoo/enterprise/blob/c4604d8b398713898b014e6073219d6202c154c3/mrp_workorder/models/mrp_workorder.py#L284-L286
The state of the work order will then be updated:
https://github.com/odoo/enterprise/blob/c4604d8b398713898b014e6073219d6202c154c3/mrp_workorder/models/mrp_workorder.py#L740"

opw-4024904

closes odoo/odoo#172587

Related: odoo/enterprise#66358
Signed-off-by: Djamel Touati (otd) <otd@odoo.com>